Protection against overflow on predictor

+++ labwork2/src/lib/predictor/predictor.cpp
@@ -18,10 +18,10 @@ short predictor::predict()
         case 1:
             return circularBuffer[0];
         case 2:
-            return (short) (2 * circularBuffer[1] - circularBuffer[0]);
+            return (short) (2 * ((int) circularBuffer[1]) - ((int) circularBuffer[0]));
         default:
-            return (short) (3 * circularBuffer[cb_ptr] - 3 * circularBuffer[(cb_ptr - 1) % 3] +
-                            circularBuffer[(cb_ptr - 2) % 3]);
+            return (short) (3 * ((int) circularBuffer[cb_ptr]) - 3 * ((int) circularBuffer[(cb_ptr - 1) % 3]) +
+                            ((int) circularBuffer[(cb_ptr - 2) % 3]));
     }
 }
